+++ tests/testthat/test-sliceFamilies.R
@@ -1,5 +1,4 @@
 library(testthat)
-library(data.table)
 library(withr)
 
 test_that("sliceFamilies creates correct output files for both mitRel = 1 and 0", {
@@ -13,8 +12,10 @@ test_that("sliceFamilies creates correct output files for both mitRel = 1 and 0"
 
   result <- com2links(
     ad_ped_matrix = ad_ped_matrix,
-    mit_ped_matrix = mit_ped_matrix, cn_ped_matrix = cn_ped_matrix,
-    writetodisk = TRUE, rel_pairs_file = "dataRelatedPairs.csv"
+    mit_ped_matrix = mit_ped_matrix,
+    cn_ped_matrix = cn_ped_matrix,
+    writetodisk = TRUE,
+    rel_pairs_file = "dataRelatedPairs.csv"
   )
   # Run function
   sliceFamilies(
@@ -41,14 +42,14 @@ test_that("sliceFamilies creates correct output files for both mitRel = 1 and 0"
   expect_gt(length(files), 0)
 
   # Check contents of a file
-  test_data <- fread("dataRelatedPairs.csv")
-  all_data <- rbindlist(lapply(files, fread), fill = TRUE)
+  test_data <- data.table::fread("dataRelatedPairs.csv")
+  all_data <- data.table::rbindlist(lapply(files, data.table::fread), fill = TRUE)
   expect_true(all(all_data$ID1 %in% test_data$ID1))
   expect_true(all(all_data$mitRel %in% c(0, 1)))
 
   # Check progress log
   expect_true(file.exists("progress.csv"))
-  progress <- fread("progress.csv")
+  progress <- data.table::fread("progress.csv")
   expect_true("start_line" %in% names(progress))
   expect_true("total_lines" %in% names(progress))